# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GlobeTalk Server API",
    "version": "1.0.0",
    "description": "OpenAPI documentation for GlobeTalk backend endpoints."
  },
  "servers": [
    { "url": "/" }
  ],
  "paths": {
...
    "/api/user/ip": {
      "post": {
        "summary": "Store user IP address",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StoreIpRequest" } } } },
        "responses": { "200": { "description": "Stored" }, "400": { "description": "Missing user ID" }, "500": { "description": "Firestore not initialized or server error" } }
      }
    },
    "/api/blockUser": {
      "post": {
        "summary": "Block a user",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["userID"], "properties": { "userID": { "type": "string" } } } } } },
        "responses": { "200": { "description": "User blocked" }, "400": { "description": "Missing userID" }, "500": { "description": "Firestore not initialized or server error" } }
      }
    },
    "/api/blocked/{userID}": {
      "get": {
        "summary": "Check if a user is blocked",
        "parameters": [ { "name": "userID", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": {
          "200": { "description": "Blocked status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BlockedStatus" } } } },
          "400": { "description": "Missing userID" },
          "500": { "description": "Firestore not initialized or server error" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StoreIpRequest": {
        "type": "object",
        "required": ["uid"],
        "properties": {
          "uid": { "type": "string" },
          "ipAddress": { "type": "string" }
        }
      },
      "BlockedStatus": {
        "type": "object",
        "properties": {
          "blocked": { "type": "boolean" },
          "source": { "type": "string" },
          "blockedAt": { "type": "number" }
        }
      },
...
    }
  }
}